bot/requirements.txt
aiohttp==3.8.3
async-tio==1.3.2
asyncpg==0.26.0
beautifulsoup4==4.11.1
dateparser==1.1.1
deep-translator==1.7.0
discord.py==2.3.2
jishaku==2.5.1
mystbin.py==5.0.0
psutil==5.4.3
python-dotenv==0.20.0
pytz==2022.1
wordhoard==1.5.1
